+++ src/app.js
@@ -1,23 +1,18 @@
 import express, {
   json
 } from 'express';
+const userRoute = require('./routes/userRoute');
 
 const app = express();
 require('dotenv').config();
-
 app.use(json())
 
 const PORT = process.env.PORT || 3000;
 
-const db = require('./database/models/index');
+app.use('/api/users', userRoute);
+// const db = require('./database/models/index');
+const { connectDB, sequelize } = require('./database/config/db');
 
-db.sequelize.sync()
-  .then(() => {
-    console.log('Synced with database');
-  })
-  .catch((err) => {
-    console.log('Error syncing with database', err);
-  });
 
 app.get('/', async (req, res) => {
   res.json({
@@ -26,4 +21,11 @@ app.get('/', async (req, res) => {
   })
 });
 
-app.listen(PORT, () => console.log(`App listening at port ${PORT}`))
+app.listen(PORT, async () => {
+  console.log(`App listening at port ${PORT}`)
+  await connectDB();
+  sequelize.sync({ force: false }).then(() => {
+    console.log("✅Synced database successfully...");
+  });
+}
+)
