+++ src/database/models/user.js
@@ -30,7 +30,7 @@ const sequelize = db.sequelize;
 
 User.init({
   id: {
-    type: DataTypes.INTEGER.UNSIGNED,
+    type: DataTypes.INTEGER,
     autoIncrement: true,
     primaryKey: true
   },
